完善送平板

+++ Assets/Scripts/Logic/LuckyBoy/FiveRoundPlay.cs
@@ -83,7 +83,7 @@ public  class FiveRoundPlay : GameMisson
 
     }
 
-    public override void NoZhuaZhong(CatchTy cat, ExtendContent voiceContent, out int delytime, out string[] contents)
+    public override void NoZhuaZhong(CatchTy cat, ExtendContent voiceContent, out float delytime, out string[] contents)
     {
         delytime = 0;
         contents = null;
