+++ Assets/Scripts/Logic/LuckyBoy/ThreeRoundPlay.cs
@@ -83,7 +83,7 @@ public  class ThreeRoundPlay : GameMisson
         }
     }
 
-    public override void NoZhuaZhong(CatchTy cat, ExtendContent voiceContent, out int delytime, out string[] contents)
+    public override void NoZhuaZhong(CatchTy cat, ExtendContent voiceContent, out float delytime, out string[] contents)
     {
         if (cat == CatchTy.NoCatch)//抖动
         {
